Support repos without hooks.yaml by reading a local config
If a repo does not have a hooks.yaml any hook ids are then looked for in
.pre-commit-hooks-config.yaml in the repo that is under test. It has the same
format as a standard hooks.yaml.

The purpose of this change is to avoid needing to mirror a linter repo just
to call it. Long term, tools could be reference by a URL or other source than
a git repo making having a mirror repo even more of a hassle.