[Bugfix][Python] Cache virtual environments by full python path
Prior to this commit, a cached virtualenv could be erroneously
re-used, causing failures in pre-commit hooks.  This is caused by
incorrect normalization in `get_default_version`, resulting in a cache
key of `"python3"`, regardless of the minor version used.

This commit updates the implementation of `get_default_version` for
python-based hooks to return `os.path.realpath(sys.executable)`.  This
ensures that the correct python version is used for running pre-commit
hooks.