Fix highlighted text not readable with some color schemes
Make the foreground (text) color be the usual background color,
when highlighting text. This is done by setting the foreground
color instead of the background color, then inverting (code 7)
to swap the foreground and background colors.

Previously, with some color schemes there was not enough contrast
between the text and highlight color.
Based on the assumption that there should be contrast between
terminal colors and the background, this should make the text
readable with a wider set of color schemes.